requirements.txt
streamlit
pandas
numpy
scikit-learn
torch
torchvision
Pillow
folium
branca
opencv-python
matplotlib
piexif
jsonschema
python-dotenv
